+++ src/neptune/internal/websockets/websockets_factory.py
@@ -16,7 +16,6 @@
 __all__ = ["WebsocketsFactory"]
 
 import threading
-from typing import Optional
 
 from requests_oauthlib import OAuth2Session
 
@@ -24,15 +23,13 @@ from neptune.internal.websockets.reconnecting_websocket import ReconnectingWebso
 
 
 class WebsocketsFactory:
-    def __init__(self, url: str, session: OAuth2Session, proxies: Optional[dict] = None):
+    def __init__(self, url: str, session: OAuth2Session):
         self._url = url
         self._session = session
-        self._proxies = proxies
 
     def create(self):
         return ReconnectingWebsocket(
             url=self._url,
             oauth2_session=self._session,
             shutdown_event=threading.Event(),
-            proxies=self._proxies,
         )