+++ src/neptune/objects/project.py
@@ -65,7 +65,6 @@ class Project(WithBackend):
             If left empty, the value of the NEPTUNE_MODE environment variable is used.
             If no value was set for the environment variable, "async" is used by default.
             Possible values are `async`, `sync`, `read-only`, and `disabled`.
-        proxies: Argument passed to HTTP calls made via the Requests library, as dictionary of strings.
 
     Returns:
         Project object that can be used to interact with the project as a whole,
@@ -107,7 +106,6 @@ class Project(WithBackend):
         *,
         api_token: Optional[str] = None,
         mode: Optional[Literal["async", "sync", "read-only", "disabled"]] = None,
-        proxies: Optional[dict] = None,
     ):
         mode = mode or Mode.READ_ONLY.value
 
@@ -120,7 +118,6 @@ class Project(WithBackend):
             project=project,
             api_token=api_token,
             mode=mode,
-            proxies=proxies,
         )
 
     def _fetch_entries(